pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rep_check"
version = "0.0.1"
description = ""
requires-python = ">=3.10"
dependencies = [
  "zarr==2.17.0",
  "numpy==1.26.4",
  "opencv-python==4.10.0.84",
  "torch==2.2.0",
  "torchvision==0.17.0",
  "mediapipe==0.10.21",
  "einops==0.8.1",
  "hydra-core==1.3.2",
  "omegaconf==2.3.0",
  "tqdm==4.67.1",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["rep_check*"]
